+++ R/regression_frame_glmmTMB.R
@@ -60,7 +60,8 @@ as_regression_frame.glmmTMB <- function(fit,
     coefs, fit, vcov, cluster, ci_level,
     test = "z", estimates = glmmTMB::fixef(fit)$cond
   )
-  coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns)
+  coefs <- .attach_ame_to_frame_coefs(coefs, fit, ci_level, show_columns,
+                                      vcov_type = vcov, cluster = cluster)
   coefs <- .attach_partial_chi2_to_frame_coefs(coefs, fit, show_columns)
   coefs <- .attach_beta_to_frame_coefs(coefs, fit, standardized, ci_level)
   info  <- .glmmTMB_info(fit,